Fixed issue with ALSA interface sometimes failing to initialise the hardware with "File descriptor in bad state" error. This is transient and the initialisation needs to be retried after a short while.
Fixed stderr message saying free of memory object not on heap. I cannot
find a definitive answer to if snd_card_get_name returns a pointer in 
the heap or not. A web search shows that it is quite common for people
to do what we did and do a free() of the pointer, but many other
implementations don't. We will assume it isn't for all drivers.

General code clean up.


git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@22737 023b2edf-31b2-4de3-b41e-bca80c47788f