Fixed ID: 3437294 "Explicitly cast the pointer in memcpyto silence clan warning", found by ojab.
Note, fix is more complicated than it seems as the "obj" parameter should really be checked for being a compatible derived class before trying to do the memcmp. The dynamic_cast does this with a static function, but the new code cannot be used with classes that are multiply inherited with PObject as command ancestor. This is generally a "bad idea" so that code was changed too.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@26772 023b2edf-31b2-4de3-b41e-bca80c47788f