Fixed notifiers broken in recent change.
Use of void * in PNotifierFunctionTemplate is really bad, but changing it exceeds backward compatibility breakage threshold. The previous change fixes it a bit, but still has the problem of multiple inheritance pointers being bogus if they are cast to void *, dependent on order of inheritance. Forgot to put that back after backing part way out of the really type safe changes.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@27436 023b2edf-31b2-4de3-b41e-bca80c47788f