Cleaned up the class hierarchy for thread synchronisation objects, derived from PSync. In particular removing PSemaphore from some derived classes.
Removed useless WillBlock() function. Between when it is called, and whatever it is you might want to do on that return value, it could change, so function is dangerous.

Added more help in tracking down deadlocks. In particular indicating which thread owns a PTimedMutex or PReadWriteMutex.

Added PWin32Handle class wrapper around WIN32 HANDLE for use in numerous places.