Merged revision(s) 28252-28254 from ptlib/branches/v2_10: Reversed out the previous sound volume changes for Windows MultiMedia API. They are too specific and do not support all devices, e.g. USB headset. Code was returned to the original, which set the master volume control.
The actual bug may have been a bogus and to 0xfffff of line.dwLineID.

Removed logarithmic scaling of Windows MultiMedia volume control.

Fixed microphone volume control for XP as it does not have a "master" volume control. This has the unfortunate side affect of only the microphone has a volume control and if the user selects line in or something it does not work. But as 99.9% of the time it is a microphone, we live with the limitation.