Merged revision(s) 30211-30212 from ptlib/branches/v2_12: Added configure option to disable use of pthread_kill() to determine if a thread has terminated. An alternate method using the /proc system is used, but this is inefficient and not universal to all Unix flavours.
Apparently, on some Linux, and/or glibc, systems, pthread_kill() will crash if the thread id is invalid.
Now, IMHO, a pthread function that is not itself thread safe is utter stupidity.
Really, what is the point of an ESRCH errno value for invalid id, if it can crash on an invalid id?
And following various blogs, the authors would rather change the Posix standard than fix it!
Complete madness.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@30213 023b2edf-31b2-4de3-b41e-bca80c47788f