Added configure support for various profiling methods: gprof, eccam, raw or manual. The latter two are "home gown" methods, and while very very primitive, they can be enough to find certain class es of problems.
The "raw" method uses the GCC -finstrument-functions option to get the execution time for every function. At the end of the run the environment variables PTLIB_RAW_PROFILING_FILENAME and PTLIB_PROFILING_FILENAME are checked and data dropped to those files.

The "manual" method uses the same internal intfrastructure as "raw" but does not use the compiler flag and you must use the macro PPROFILE_FUNCTION() at the top of any functions you wish to profi
le.