Revert "Added glibc CXX/STL debugging flag when compiling in debug mode"
This reverts commit a537de725898c2a028fbbf2db07a8395756ee568.

I had to revert this due to some very, very strange interaction with the video plugins. While all documentation says that different translation units can use, or not use, debug iterators, it seems this is not exactly the case if you have some common code that is in those two translation units (one off, one on) despite the fact that the common code is marked "static" so there should be two complete copies of it. Tried several techniques to fix, e.g. using namespaces, templates, but to no avail, the "off" version ends up calling the debug iterator on a non debug iterator container (std::map) and it crashes.