Fixed possible race condition closing PIndirectChannel.
There is a particular code pattern where a thread is reading a PIndirectChannel, the base channel must be closed to break the read block. This was always done via code similar to GetBaseReadChannel()->Close(). However, this opens up a small window in which that base channel could have been deleted, via PIndirectChannel::Close(), before Close() is called on it. Doing the base channel Close() within the channelPointerMutex should plug that hole.